% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIEN_public_AP.R
\name{BIEN.gis.box}
\alias{BIEN.gis.box}
\title{Occurrence records from from lat/long Bounding box}
\usage{
BIEN.gis.box(min.lat, max.lat, min.long, max.long, cultivated = FALSE,
  only.new.world = TRUE)
}
\arguments{
\item{min.lat}{A number -90 to 90}

\item{max.lat}{A number -90 to 90 that is greater than min.lat}

\item{min.long}{A number -180 to 180}

\item{max.long}{A number -180 to 180 that is greater than min.long}

\item{cultivated}{cultivated = Return cultivated records as well?  Default is FALSE.}

\item{only.new.world}{only.new.world = Return only records from the New World?  Default is TRUE.}
}
\description{
Occurrence records from from lat/long Bounding box
}
\examples{
output_test<-BIEN.gis.box(min.lat = 32,max.lat = 33,min.long = -114,max.long = -113)
}
\author{
Brian Maitner

Will only do one box at a time, specified by min/max lat and long
}

