% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIEN_public_AP.R
\name{BIEN.gis.country}
\alias{BIEN.gis.country}
\title{Occurrences from Country}
\usage{
BIEN.gis.country(country, cultivated = FALSE, only.new.world = TRUE)
}
\arguments{
\item{country}{Accepts a single Country or a vector}

\item{cultivated}{cultivated = Return cultivated records as well?  Default is FALSE.}

\item{only.new.world}{only.new.world = Return only records from the New World?  Default is true.  Probably not very useful to change at the moment.}
}
\description{
Occurrences from Country
}
\note{
Insanely slow
}
\examples{
BIEN.gis.state("Mexico")
state_vector<-c("Mexico","Canada")
BIEN.gis.state(state_vector)
}
\author{
Brian Maitner
}

