% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIEN_public_AP.R
\name{BIEN.gis.family}
\alias{BIEN.gis.family}
\title{Occurrences from Family}
\usage{
BIEN.gis.family(family, cultivated = FALSE, only.new.world = TRUE)
}
\arguments{
\item{family}{Accepts a single Family or a vector of Families}

\item{cultivated}{cultivated = Return cultivated records as well?  Default is FALSE.}

\item{only.new.world}{only.new.world = Return only records from the New World?  Default is TRUE.}
}
\description{
Occurrences from Family
}
\examples{
BIEN.gis.family("Theaceae")
family_vector<-c("Theaceae","Ericaceae")
BIEN.gis.family(family_vector)
}
\author{
Brian Maitner
}

