% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIEN_public_AP.R
\name{BIEN.gis.genus}
\alias{BIEN.gis.genus}
\title{Occurrences from Genus}
\usage{
BIEN.gis.genus(genus, cultivated = FALSE, only.new.world = TRUE)
}
\arguments{
\item{genus}{Accepts a single Genus or a vector of Genera}

\item{cultivated}{cultivated = Return cultivated records as well?  Default is FALSE.}

\item{only.new.world}{only.new.world = Return only records from the New World?  Default is TRUE.}
}
\description{
Occurrences from Genus
}
\examples{
BIEN.gis.genus("Abutilon")
genus_vector<-c("Abutilon","Abronia")
BIEN.gis.genus(genus_vector)
BIEN.gis.genus(genus = "Abutilon",cultivated = TRUE,only.new.world = FALSE)
#returns all records for the genus Abutilon, including cultivates and Old World records.
}
\author{
Brian Maitner
}

