% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIEN_public_AP.R
\name{BIEN.gis.species}
\alias{BIEN.gis.species}
\title{Lat/Long from Species Names(s)}
\usage{
BIEN.gis.species(species, cultivated = FALSE, only.new.world = TRUE)
}
\arguments{
\item{species}{species here can be either a single species or a vector of species}

\item{cultivated}{cultivated = Return cultivated records as well?  Default is FALSE.}

\item{only.new.world}{only.new.world = Return only records from the New World?  Default is TRUE.}
}
\description{
Lat/Long from Species Names(s)
}
\examples{
BIEN.gis.species("Abies amabilis")
species_vector<-c("Abies amabilis", "Acer nigrum")
BIEN.gis.species(species_vector)
}
\author{
Brian Maitner
}

