% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIEN_public_AP.R
\name{BIEN.gis.state}
\alias{BIEN.gis.state}
\title{Occurrences from State/Province}
\usage{
BIEN.gis.state(state, cultivated = FALSE, only.new.world = TRUE)
}
\arguments{
\item{state}{Accepts a single State or a vector}

\item{cultivated}{cultivated = Return cultivated records as well?  Default is FALSE.}

\item{only.new.world}{only.new.world = Return only records from the New World?  Default is TRUE.  Probably not very useful to change at the moment.}
}
\description{
Occurrences from State/Province
}
\note{
Insanely slow
}
\examples{
BIEN.gis.state("Rhode Island")
state_vector<-c("Rhode Island","Maryland")
BIEN.gis.state(state_vector)
}
\author{
Brian Maitner
}

