% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIEN_public_AP.R
\name{BIEN.list.country}
\alias{BIEN.list.country}
\title{Species list from country/countries}
\usage{
BIEN.list.country(country, cultivated = FALSE, only.new.world = TRUE)
}
\arguments{
\item{country}{Accepts single countries or vectors}

\item{cultivated}{cultivated = Return cultivated records as well?  Default is FALSE.}

\item{only.new.world}{only.new.world = Return only records from the New World?  Default is TRUE.}
}
\description{
Species list from country/countries
}
\examples{
BIEN.list.country("Canada")
country_vector<-c("Canada","United States")
BIEN.list.country(country_vector)
}
\author{
Brian Maitner
}

