% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIEN_public_AP.R
\name{BIEN.list.state}
\alias{BIEN.list.state}
\title{Species list from state(s)}
\usage{
BIEN.list.state(state, cultivated = FALSE, only.new.world = TRUE)
}
\arguments{
\item{state}{Accepts single countries or vectors}

\item{cultivated}{cultivated = Return cultivated records as well?  Default is FALSE.}

\item{only.new.world}{only.new.world = Return only records from the New World?  Default is TRUE.}
}
\description{
Species list from state(s)
}
\examples{
BIEN.list.state("Michigan")
state_vector<-c("Michigan","Arizona")
BIEN.list.state(state_vector)
}
\author{
Brian Maitner
}

