% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIEN_public_AP.R
\name{BIEN.ranges.species}
\alias{BIEN.ranges.species}
\title{BIEN ranges function}
\usage{
BIEN.ranges.species(species, directory = NULL, matched = TRUE,
  match_names_only = FALSE)
}
\arguments{
\item{species}{species as either a single species or a vector of species}

\item{directory}{the directory where you want the range shapefiles saved. Default is your current working directory}

\item{matched}{returns information on whether range maps were found for each species submitted}

\item{match_names_only}{Do you want to check whether the maps you're looking for exist without downloading them?}
}
\description{
BIEN ranges function
}
\note{
The code marked as not run is only marked that way because the directories are not initialized.
}
\examples{
species_vector<-c("Abies_lasiocarpa","Abies_amabilis")
BIEN.ranges.species(species_vector)
BIEN.ranges.species("Abies_lasiocarpa")
\dontrun{
BIEN.ranges.species("Abies_lasiocarpa","/Users/Alton/Documents/GitHub/BIEN3Project")
}
}
\author{
Brian Maitner
}

