% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIEN_public_AP.R
\name{BIEN.trait.family}
\alias{BIEN.trait.family}
\title{Traits from family/families}
\usage{
BIEN.trait.family(family)
}
\arguments{
\item{genus}{Accepts a single family or a vector}
}
\description{
Traits from family/families
}
\examples{
#see BIEN.trait.species
}
\author{
Brian Maitner
}

