% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIEN_public_AP.R
\name{BIEN.trait.genus}
\alias{BIEN.trait.genus}
\title{Traits from species}
\usage{
BIEN.trait.genus(genus)
}
\arguments{
\item{genus}{Accepts a single genus or a vector}
}
\description{
Traits from species
}
\examples{
#see BIEN.trait.species
}
\author{
Brian Maitner
}

