% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIEN_public_AP.R
\name{BIEN.trait.list}
\alias{BIEN.trait.list}
\title{List all traits}
\usage{
BIEN.trait.list()
}
\value{
This function lists all currently available types of trait data. This is especially useful if you want to figure out what your trait of interest is titled in this database.
}
\description{
List all traits
}
\author{
Brian Maitner
}

