% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIEN_public_AP.R
\name{BIEN.trait.species}
\alias{BIEN.trait.species}
\title{Traits from species}
\usage{
BIEN.trait.species(species)
}
\arguments{
\item{species}{Accepts a single species or a vector}
}
\description{
Traits from species
}
\examples{
BIEN.trait.species("Poa annua")
species_vector<-c("Poa annua","Juncus trifidus")
BIEN.trait.species(species_vector)
}
\author{
Brian Maitner
}

