% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIEN_public_AP.R
\name{BIEN.trait.trait}
\alias{BIEN.trait.trait}
\title{Traits from trait name}
\usage{
BIEN.trait.trait(trait)
}
\arguments{
\item{trait}{Accepts a single trait or a vector}
}
\description{
Traits from trait name
}
\examples{
BIEN.trait.trait("Height")
trait_vector<-c("Height", "Leaf dry mass")
BIEN.trait.trait(trait_vector)
}
\author{
Brian Maitner
}

