% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIEN_public_AP.R
\name{BIEN.trait.traitbyfamily}
\alias{BIEN.trait.traitbyfamily}
\title{Specific traits from family/families}
\usage{
BIEN.trait.traitbyfamily(trait, family)
}
\arguments{
\item{trait}{Accepts a single trait or a vector}

\item{family}{Accepts a single family or a vector}
}
\value{
This function downloads specific trait data for given family/families
}
\description{
Specific traits from family/families
}
\examples{
#see BIEN.trait.traitbyspecies
}
\author{
Brian Maitner
}

