% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIEN_public_AP.R
\name{BIEN.trait.traitbygenus}
\alias{BIEN.trait.traitbygenus}
\title{Specific traits from genus/genera}
\usage{
BIEN.trait.traitbygenus(trait, genus)
}
\arguments{
\item{trait}{Accepts a single trait or a vector}

\item{genus}{Accepts a single genus or a vector}
}
\description{
Specific traits from genus/genera
}
\examples{
#see BIEN.trait.traitbyspecies
}
\author{
Brian Maitner

This function downloads specific trait data for given genus/genera
}

