% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIEN_public_AP.R
\name{BIEN.trait.traitbyspecies}
\alias{BIEN.trait.traitbyspecies}
\title{Specific traits from species}
\usage{
BIEN.trait.traitbyspecies(trait, species)
}
\arguments{
\item{trait}{Accepts a single trait or a vector}

\item{species}{Accepts a single species or a vector}
}
\description{
Specific traits from species
}
\examples{
BIEN.trait.traitbyspecies(trait = "Height", species = "Carex capitata")
trait_vector<-c("Height", "Leaf dry mass")
species_vector<-c("Carex capitata","Betula nana")
BIEN.trait.traitbyspecies(trait=trait_vector,species=species_vector)
}
\author{
Brian Maitner

This function downloads trait data for given species
}

